requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
Deprecated==1.2.10
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
PyGithub==1.51
PyJWT==1.7.1
pylint==2.5.3
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
